Test inline functions with DistributedQueryRunner
